requirements.txt
numpy==1.19.5
tensorflow==2.4.1
matplotlib==3.3.4
Pillow==8.1.0
albumentations==0.5.2
seaborn==0.11.1
keract==4.4.0
opencv-python==4.5.1.48
Keras==2.4.3
sklearn==0.0
scikit-learn==0.24.2
plotly==5.2.2
pandas==1.2.1